Refactor model package
