+++ src/main/java/com/mangalaxy/mango/model/entity/Education.java
@@ -1,5 +1,7 @@
-package com.mangalaxy.mango.entity;
+package com.mangalaxy.mango.model.entity;
 
+import com.mangalaxy.mango.model.ActivityPeriod;
+import com.mangalaxy.mango.model.Degree;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
