+++ src/main/java/com/mangalaxy/mango/model/entity/Experience.java
@@ -1,5 +1,6 @@
-package com.mangalaxy.mango.entity;
+package com.mangalaxy.mango.model.entity;
 
+import com.mangalaxy.mango.model.ActivityPeriod;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
