+++ src/main/java/com/mangalaxy/mango/repository/EmployerRepository.java
@@ -1,6 +1,6 @@
 package com.mangalaxy.mango.repository;
 
-import com.mangalaxy.mango.entity.Employer;
+import com.mangalaxy.mango.model.entity.Employer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EmployerRepository extends JpaRepository<Employer, Long> {
