+++ src/main/java/com/mangalaxy/mango/repository/LocationRepository.java
@@ -1,6 +1,6 @@
 package com.mangalaxy.mango.repository;
 
-import com.mangalaxy.mango.entity.Location;
+import com.mangalaxy.mango.model.entity.Location;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LocationRepository extends JpaRepository<Location, Long> {
