+++ src/main/java/com/mangalaxy/mango/repository/ProfileRepository.java
@@ -1,6 +1,6 @@
 package com.mangalaxy.mango.repository;
 
-import com.mangalaxy.mango.entity.Profile;
+import com.mangalaxy.mango.model.entity.Profile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProfileRepository extends JpaRepository<Profile, Long> {
