+++ src/main/java/com/mangalaxy/mango/repository/SkillRepository.java
@@ -1,6 +1,6 @@
 package com.mangalaxy.mango.repository;
 
-import com.mangalaxy.mango.entity.Skill;
+import com.mangalaxy.mango.model.entity.Skill;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SkillRepository extends JpaRepository<Skill, Long> {
