+++ src/main/java/com/mangalaxy/mango/repository/TalentRepository.java
@@ -1,6 +1,6 @@
 package com.mangalaxy.mango.repository;
 
-import com.mangalaxy.mango.entity.Talent;
+import com.mangalaxy.mango.model.entity.Talent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
