+++ src/main/resources/application.properties
@@ -15,6 +15,6 @@ spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 # only for development mode, change for production stage
 spring.jpa.generate-ddl=true
 spring.jpa.hibernate.ddl-auto=update
-logging.level.org.hibernate.SQL=DEBUG
-logging.level.org.hibernate.type=TRACE
+#logging.level.org.hibernate.SQL=DEBUG
+#logging.level.org.hibernate.type=TRACE
 logging.level.web=DEBUG
